requirements.txt
asgiref==3.3.1
Django==3.2a1
pytz>=2021.3
sqlparse==0.4.1
djangorestframework==3.11
django-redis==5.2.0
pandas==1.5.3
celery==5.2.7
openpyxl==3.0.10